Refactor Connecting process Removes unnecessary verbose activity + viewModel + fragment by a standalone activity
