+++ fe2-android/app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
     android:roundIcon="@mipmap/ic_launcher_round"
     android:supportsRtl="true"
     android:theme="@style/AppTheme">
-    <activity android:name=".ui.home.connecting.ConnectingActivity" />
+    <activity android:name=".ui.home.ConnectingActivity" />
     <activity
       android:name=".ui.home.HomeActivity"
       android:exported="true">
