+++ fe2-android/app/src/main/res/layout/connecting_activity.xml
@@ -1,9 +1,53 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-  xmlns:tools="http://schemas.android.com/tools"
+  xmlns:app="http://schemas.android.com/apk/res-auto"
   android:id="@+id/fragment_container_connecting"
   android:layout_width="match_parent"
-  android:layout_height="match_parent"
-  tools:context=".ui.home.connecting.ConnectingActivity">
+  android:layout_height="match_parent">
 
+  <Button
+    android:id="@+id/button_cancel_connecting"
+    style="@style/button_style"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:layout_margin="@dimen/lone_bottom_button_top_margin"
+    android:backgroundTint="@color/gray"
+    android:text="@string/button_cancel"
+    app:layout_constraintBottom_toBottomOf="parent"
+    app:layout_constraintEnd_toEndOf="parent"
+    app:layout_constraintStart_toStartOf="parent" />
+
+  <LinearLayout
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:gravity="center"
+    android:orientation="vertical">
+
+    <RelativeLayout
+      android:id="@+id/loadingPanel"
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content"
+      android:gravity="center">
+
+      <ProgressBar
+        android:id="@+id/connecting_progress_bar"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:indeterminate="true" />
+    </RelativeLayout>
+
+    <TextView
+      android:id="@+id/connecting_lao"
+      style="@style/explication_welcome_style"
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content" />
+
+    <TextView
+      android:id="@+id/connecting_text"
+      style="@style/explication_welcome_style"
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content"
+      android:text="@string/connecting" />
+
+  </LinearLayout>
 </androidx.constraintlayout.widget.ConstraintLayout>
