--- fe2-android/app/src/main/res/layout/connecting_fragment.xml
@@ -1,66 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:app="http://schemas.android.com/apk/res-auto"
-  xmlns:android="http://schemas.android.com/apk/res/android">
-
-  <data>
-
-    <import type="android.view.View" />
-
-    <variable
-      name="view_model"
-      type="com.github.dedis.popstellar.ui.home.connecting.ConnectingViewModel" />
-  </data>
-
-  <androidx.constraintlayout.widget.ConstraintLayout
-    android:id="@+id/fragment_connecting"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <LinearLayout
-      android:layout_width="match_parent"
-      android:layout_height="match_parent"
-      android:gravity="center"
-      android:orientation="vertical">
-
-      <TextView
-        android:id="@+id/connecting_text"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/connecting"
-        style="@style/explication_welcome_style" />
-
-      <TextView
-        android:id="@+id/connecting_lao"
-        android:text="@{view_model.getConnectingLao}"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        style="@style/explication_welcome_style" />
-
-      <RelativeLayout
-        android:id="@+id/loadingPanel"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="center">
-
-        <ProgressBar
-          android:id="@+id/connecting_progress_bar"
-          android:layout_width="wrap_content"
-          android:layout_height="wrap_content"
-          android:indeterminate="true" />
-      </RelativeLayout>
-
-    </LinearLayout>
-
-    <Button
-      android:id="@+id/button_cancel_connecting"
-      style="@style/button_style"
-      android:layout_width="wrap_content"
-      android:layout_height="wrap_content"
-      android:text="@string/button_cancel"
-      app:layout_constraintBottom_toBottomOf="parent"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintStart_toStartOf="parent"
-      android:layout_margin="@dimen/lone_bottom_button_top_margin"
-      android:backgroundTint="@color/gray" />
-  </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
