+++ fe2-android/app/src/test/java/com/github/dedis/popstellar/ui/pages/home/ConnectingPageObject.java
@@ -1,6 +1,5 @@
 package com.github.dedis.popstellar.ui.pages.home;
 
-import androidx.annotation.IdRes;
 import androidx.test.espresso.ViewInteraction;
 
 import com.github.dedis.popstellar.R;
@@ -9,11 +8,6 @@ import static androidx.test.espresso.Espresso.onView;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 
 public class ConnectingPageObject {
-
-  @IdRes
-  public static int connectingFragmentContainerId() {
-    return R.id.fragment_container_connecting;
-  }
 
   public static ViewInteraction connectingText() {
     return onView(withId(R.id.connecting_text));
